Use better naming convention for Payment Methods/Gateways
